backend_django/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
django-filter==23.3
celery==5.3.4
redis==5.0.1
requests==2.31.0
gunicorn==21.2.0

# Optional PostgreSQL support (uncomment if needed)
# psycopg2-binary==2.9.7